api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==4.1.2
blinker==1.6.2
cachetools==5.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
configparser==5.3.0
cryptography==42.0.3
csrf==0.1b1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.5
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
executing==1.2.0
fastapi==0.109.2
fastapi-csrf-protect==0.2.2
fastapi-mail==1.4.1
fastapi-pagination==0.12.4
ffmpeg-python==0.2.0
filelock==3.12.4
fpdf==1.7.2
frozenlist==1.3.3
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httptools==0.5.0
httpx==0.26.0
idna==3.6
ipinfo==4.4.2
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jose==1.0.0
jupyter_client==8.3.1
jupyter_core==5.3.1
keyboard==0.13.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
MetaTrader5==5.0.45
multidict==6.0.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
nest-asyncio==1.5.8
numpy==1.24.1
orjson==3.8.12
packaging==23.1
pandas==2.1.1
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pipenv==2023.10.3
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.6.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
pyfiglet==0.8.post1
Pygments==2.16.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.7
pywin32==306
PyYAML==6.0
pyzmq==25.1.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
send_email==201212
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.36.3
Telethon==1.26.1
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.9.0
tzdata==2022.1
ujson==5.7.0
urllib3==2.2.1
uvicorn==0.27.1
vboxapi==1.0
virtualenv==20.24.5
watchfiles==0.19.0
wcwidth==0.2.6
websockets==10.3
yarl==1.9.2
